Version 1.1
Adds ability to verify bulk motor direction on very first time startup before storing result in EEPROM. Necessary because motor cable polarity is less consistent than desired, will be configured prior to shipping so users will only ever need this when they complete a software update.

It will rotate the bulk motor, attempting to make it go Counter-Clockwise. Push and hold the left or right button based on whether it's spinning to the left (CCW) or to the right (CW) when prompted. If you make a mistake, toggle the enable button during the confirmation period (when it asks if it's spinning correctly now) and it will cancel so you can re-try during the next power cycle.